moving metadata namespace up a level
