+++ mojap_metadata/__init__.py
@@ -1,0 +1,1 @@
+from .metadata.metadata import Metadata  # noqa: 401
