--- mojap_metadata/metadata/__init__.py
@@ -1,1 +1,0 @@
-from .metadata import Metadata  # noqa: 401
